Add files via upload
This program downloads the open data of COVID-19 from the website of the Ministry of Health, Labour and Welfare, the Government of Japan, and draws the graph of the record of the total cases, the total discharged individuals and the active cases. I am very sorry to say that the deaths are included in the active cases in order to compare with the SIR model. I would like to express my sincere condolence of the victims and their family members.